1.机械参数页面中ABC轴安全点“机械长度”修改为“最大旋转” 2.相同界面上“每转距离”改为“电机每圈”，并将所有单位改为“度” 3.安全点页面中ABC轴页面去掉“最大移动”一栏，只保留“横行安全范围（大于）（小于）”

+++ mainframe.cpp
@@ -598,7 +598,7 @@ void MainFrame::StatusRefreshed()
         isY2PosChanged_ = true;
     }
 
-    pos = virtualHost->HostStatus(ICVirtualHost::APos).toInt();
+    pos = (qint16)virtualHost->HostStatus(ICVirtualHost::APos).toInt();
     if(pos != oldAPos_)
     {
         oldAPos_ = pos;
@@ -607,7 +607,7 @@ void MainFrame::StatusRefreshed()
         isAPosChanged_ = true;
     }
 
-    pos = virtualHost->HostStatus(ICVirtualHost::BPos).toInt();
+    pos = (qint16)virtualHost->HostStatus(ICVirtualHost::BPos).toInt();
     if(pos != oldBPos_)
     {
         oldBPos_ = pos;
@@ -616,7 +616,7 @@ void MainFrame::StatusRefreshed()
         isBPosChanged_ = true;
     }
 
-    pos = virtualHost->HostStatus(ICVirtualHost::CPos).toInt();
+    pos = (qint16)virtualHost->HostStatus(ICVirtualHost::CPos).toInt();
     if(pos != oldCPos_)
     {
         oldCPos_ = pos;
